nif-master/requirements.txt
absl-py==1.4.0
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
Brotli==1.0.9
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.3
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.12.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lazy_loader==0.2
lit==16.0.3
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
packaging==23.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==4.23.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytorch-msssim==0.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==26.2.0
referencing==0.35.1
requests==2.30.0
requests-oauthlib==1.3.1
rpds-py==0.20.0
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tifffile==2023.4.12
tinycss2==1.3.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4.1
traitlets==5.14.3
triton==2.0.0
typing_extensions==4.5.0
urllib3==2.0.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.3.4
yarg==0.1.9
zipp==3.15.0
